Move test to resource_processor_test.go
